requirements.txt
asgiref==3.10.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.8
django-cors-headers==4.9.0
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
gprof2dot==2025.4.14
h11==0.16.0
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
qrcode==8.2
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
